% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_grandmasters.R
\docType{data}
\name{chess_grandmasters}
\alias{chess_grandmasters}
\title{Chess Grand Masters  Data}
\format{
A tibble with 2044 rows and 12 columns:
\itemize{
\item \code{name}: Name of the chess player.
\item \code{fide_id}: Fide id of the player(NA if honorary).
\item \code{date_of_birth}: Date of birth of the player.
\item \code{birthplace}: Place of birth of the player.
\item \code{died}: Date of death of the player (NA if alive).
\item \code{title_year}: Year when the player achieved a title.
\item \code{federation}: Chess federation the player is associated with.
\item \code{gender}: Gender of the player.
\item \code{notes}: Additional notes about the player.
\item \code{links}: Fide id link of the player.
\item \code{latitude}: Latitude of the player's birthplace.
\item \code{longitude}: Longitude of the player's birthplace.
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_chess_grandmasters}
}
\usage{
chess_grandmasters
}
\description{
A dataset containing information about chess grandmasters in the world.
}
\examples{
# Load the dataset
data("chess_grandmasters")

# View the first few rows of the dataset
head(chess_grandmasters)
}
\keyword{datasets}
