% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChessGMsdata-package.R
\docType{package}
\name{ChessGMsdata-package}
\alias{ChessGMsdata}
\alias{ChessGMsdata-package}
\title{ChessGMsdata: Package With Chess Grandmasters data}
\description{
This package offers a comprehensive dataset on chess grandmasters.
}
\author{
\strong{Maintainer}: Jehangeer Aswani \email{aswanijehangeer@gmail.com} (\href{https://orcid.org/0009-0008-6693-5227}{ORCID})

}
\keyword{internal}
